Make form disabled until it is valid
[MAILPOET-3222]
